+++ plugins/main/public/components/common/data-source/pattern/fim/data-source.ts
@@ -15,11 +15,11 @@ export class FIMFilesStatesDataSource extends PatternDataSource {
   }
 
   getFixedFilters(): tFilter[] {
-    return [...this.getClusterManagerFilters(), ...super.getFixedFilters()];
+    return [...this.getClusterFilters(), ...super.getFixedFilters()];
   }
 
-  getClusterManagerFilters() {
-    return PatternDataSourceFilterManager.getClusterManagerFilters(
+  getClusterFilters() {
+    return PatternDataSourceFilterManager.getClusterFilters(
       this.id,
       DATA_SOURCE_FILTER_CONTROLLED_CLUSTER_MANAGER,
       VULNERABILITY_IMPLICIT_CLUSTER_MODE_FILTER,
@@ -33,11 +33,11 @@ export class FIMRegistryKeysStatesDataSource extends PatternDataSource {
   }
 
   getFixedFilters(): tFilter[] {
-    return [...this.getClusterManagerFilters(), ...super.getFixedFilters()];
+    return [...this.getClusterFilters(), ...super.getFixedFilters()];
   }
 
-  getClusterManagerFilters() {
-    return PatternDataSourceFilterManager.getClusterManagerFilters(
+  getClusterFilters() {
+    return PatternDataSourceFilterManager.getClusterFilters(
       this.id,
       DATA_SOURCE_FILTER_CONTROLLED_CLUSTER_MANAGER,
       VULNERABILITY_IMPLICIT_CLUSTER_MODE_FILTER,
@@ -51,11 +51,11 @@ export class FIMRegistryValuesStatesDataSource extends PatternDataSource {
   }
 
   getFixedFilters(): tFilter[] {
-    return [...this.getClusterManagerFilters(), ...super.getFixedFilters()];
+    return [...this.getClusterFilters(), ...super.getFixedFilters()];
   }
 
-  getClusterManagerFilters() {
-    return PatternDataSourceFilterManager.getClusterManagerFilters(
+  getClusterFilters() {
+    return PatternDataSourceFilterManager.getClusterFilters(
       this.id,
       DATA_SOURCE_FILTER_CONTROLLED_CLUSTER_MANAGER,
       VULNERABILITY_IMPLICIT_CLUSTER_MODE_FILTER,